WQS-115: Refactor Playwright test structure and update README
